pretty much everything validated. Marketplace forms still need a price field

+++ lib/widgets/weather/current_weather_item.dart
@@ -24,7 +24,7 @@ class Weather {
 
   factory Weather.fromJson(Map<String, dynamic> json) {
     return Weather(
-      temp: json['main']['temp'],
+      temp: json['main']['temp'].toDouble(),
       weather: json['weather'][0]['description'],
       icon: json['weather'][0]['icon'],
     );
